Refactor: Product Listing page
