+++ src/helpers/index.js
@@ -1,3 +1,7 @@
-export { sortData, filterProductsUptoPriceRange, categorizeData, rateData, getOnlyFastDeliveryData, getStockData, searchData } from "./filterHelpers";
+export { getCurrentPageProducts, getFilteredProducts } from "./filterHelpers";
 export { checkIfAllInputsAreNotEmpty } from "./authHelpers";
-export { isItemInArrayOfObjects, deepCloneObject, capitalizeFirstWordOfString } from "./utilityHelpers";
+export {
+  isItemInArrayOfObjects,
+  deepCloneObject,
+  capitalizeFirstWordOfString,
+} from "./utilityHelpers";
