+++ src/services/index.js
@@ -11,3 +11,4 @@ export { addNewAddressService } from "./addressServices/addNewAddress.service";
 export { editAddressService } from "./addressServices/editAddress.service";
 export { deleteAddressService } from "./addressServices/deleteAddress.service";
 export { getAddressesService } from "./getDataServices/getAddresses.service";
+export { getAllProductsService } from "./getDataServices/getAllProducts.service";
